春耕0.4 螢幕長亮，新增Arith(能夠精準計算浮點數)，改善Y軸距離計算精度

+++ app/src/main/java/com/boss/vestibularsystemdetection/backsupportdetection/measure/CompareActivity.java
@@ -41,7 +41,7 @@ public class CompareActivity extends AppCompatActivity {
     private Runnable delayToCalculate = new Runnable() {
         @Override
         public void run() {
-            List<Integer> distance = mTrajectoryView.getDistance();
+            List<Double> distance = mTrajectoryView.getDistance();
             int distance_sum = calcaulateData(distance);
             tv11.setText(distance_sum+"");
 
@@ -52,10 +52,10 @@ public class CompareActivity extends AppCompatActivity {
         }
     };
 
-    private int calcaulateData(List<Integer> list){
+    private int calcaulateData(List<Double> list){
         int sum = 0;
-        for(int l : list){
-            sum += Math.abs(l);
+        for(double l : list){
+            sum += Math.abs((int) l);
         }
         return sum;
     }
